refactoring to introduce twitter package
