+++ go.mod
@@ -1,11 +1,12 @@
-module github.com/pankona/oidc-test
+module github.com/pankona/hashira-auth
 
 require (
-	github.com/coreos/go-oidc v2.0.0+incompatible
-	github.com/pquerna/cachecontrol v0.0.0-20180517163645-1555304b9b35 // indirect
-	github.com/satori/go.uuid v1.2.0
-	golang.org/x/crypto v0.0.0-20181015023909-0c41d7ab0a0e // indirect
-	golang.org/x/net v0.0.0-20181023162649-9b4f9f5ad519
-	golang.org/x/oauth2 v0.0.0-20181017192945-9dcd33a902f4
-	gopkg.in/square/go-jose.v2 v2.1.9 // indirect
+	github.com/ChimeraCoder/anaconda v2.0.0+incompatible
+	github.com/ChimeraCoder/tokenbucket v0.0.0-20131201223612-c5a927568de7 // indirect
+	github.com/azr/backoff v0.0.0-20160115115103-53511d3c7330 // indirect
+	github.com/dustin/go-jsonpointer v0.0.0-20160814072949-ba0abeacc3dc // indirect
+	github.com/dustin/gojson v0.0.0-20160307161227-2e71ec9dd5ad // indirect
+	github.com/garyburd/go-oauth v0.0.0-20180319155456-bca2e7f09a17
+	golang.org/x/net v0.0.0-20181102091132-c10e9556a7bc
+	golang.org/x/oauth2 v0.0.0-20181102003913-e0f2c55a7fc7
 )
